relocate the selction component
